Some changes

+++ UserInterface/Controllers/AdminsController.cs
@@ -93,7 +93,7 @@ namespace UserInterface.Controllers
             OrderConfirmed orderConfirm = new OrderConfirmed()
             {
                 OrderConfirmedId = order.OrderConfirmedId,
-                ProductId = order.ProductId,
+                ProductId = (int)order.ProductId,
                 CustomerId = order.CustomerId,
                 CustomerName = order.CustomerName,
                 CustomerAddress = order.CustomerAddress,
